+++ crates/cairo-lang-language-server/src/lsp/capabilities/client.rs
