+++ crates/cairo-lang-language-server/src/lsp/capabilities/mod.rs
@@ -1,0 +1,2 @@
+pub mod client;
+pub mod server;
